requirements.txt
pika~=1.1.0
uvicorn~=0.22.0
python-dotenv~=0.19.2
fastapi~=0.97.0
openai~=0.27.9
slack_bolt~=1.18.0
tiktoken~=0.4.0
typing~=3.7.4.3